updates to work fast on Michael's data
- changed the logic of the filepaths to allow multiple animal names.
Local copy merges experiments under the first animal name specified.

- registration works simultaneously on all planes so that each tiff can
be deleted after processing. Requirement on hard drive reduced to total
size of acquired data. This is the best we can do because the bin files
need to be stored at this size anyway.

- fixed registration instability issue: it was related to the local
center of mass using negative values of the correlation, which could
cancel out with the peak and result in a near-zero mass which was used
in a denominator.

- fixed ROI detection to split all ROIs into connected pieces, not just
the ones satisfying criteria.

- added to each region's stat the number of its parent, so that
statistics of the parent can be computed post-hoc.

-