added back tiff file writing
writes tiff files in 2000 frame chunks.

Added parameter nimgbegend which is the number of frames at the
beginning and end of each experiment to average frames over. Result is
saved in ops.mimg_beg and ops.mimg_end. This can be used to determine
whether there was drift during the experiment.