Bugfixes and new translate_movie function
Bugfix to registration_offsets on small batches

Added preliminary translate_movie function ultimately intended to
replace register_movie.
- It can return the valid x and y indices over the whole movie given the
translations to help with clipping the wrapped edges.
- Works on batches of frames to help it run a bit faster, and you can
safely run it on as many frames as you can fit in your system RAM.
- Defaults to returning translated movie with same precision input
movie. Do we need to be able to specify it?
- Uses subpixel parameter to round translations appropriately. I suggest
removing this function from registration_offsets so that always returns
the most accurate offsets, which can be rounded later (during
translation) if desired.